requirements.txt
django-import-export==2.0.2
certifi==2019.11.28
dj-database-url==0.5.0
Django==2.1
gunicorn==19.9.0
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
whitenoise==4.1.4
wincertstore==0.2
mysqlclient==1.4.6
schedule==0.6.0
djangorestframework==3.11.0